backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for Luis Martinez portfolio"
requires-python = ">=3.10"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "psycopg[binary]",
    "pydantic",
    "apscheduler",
    "python-dotenv",
    "alembic",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "black", "flake8", "isort"]
